% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ar1.R
\name{ar1_Q}
\alias{ar1_Q}
\title{Compute the precision matrix for an AR(1) process}
\usage{
ar1_Q(n_times, rho)
}
\arguments{
\item{n_times}{Number of time periods}

\item{rho}{The autocorrelation parameter, between -1 and 1}
}
\description{
Computes the sparse precision matrix for an AR(1) process. If n_times is 1,
just returns the 1x1 identity matrix.
}
