% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process-model.R
\name{aggregate_flux}
\alias{aggregate_flux}
\title{Aggregate monthly fluxes according to a constraint}
\usage{
aggregate_flux(model, filter_expr, parameters = model)
}
\description{
Aggregate monthly fluxes according to a constraint
}
